requirements.txt
python-telegram-bot==20.6
firebase-admin
requests
aiohttp
python-dotenv
